hw 6 finish

+++ src/services/auth.js
@@ -1,40 +1,45 @@
 import { randomBytes } from 'crypto';
 import bcrypt from 'bcrypt';
-import createHttpError from 'http-errors';
-
-import { FIFTEEN_MINUTES, THIRTY_DAYS } from '../constants/index.js';
-import { SessionsCollection } from '../db/models/session.js';
-
 import { UsersCollection } from '../db/models/user.js';
+import createHttpError from 'http-errors';
+import { SessionsCollection } from '../db/models/session.js';
+import {
+  FIFTEEN_MINUTES,
+  SMTP,
+  TEMPLATES_DIR,
+  THIRTY_DAYS,
+} from '../constants/index.js';
+import jwt from 'jsonwebtoken';
+import { getEnvVar } from '../utils/getEnvVar.js';
+import { sendEmail } from '../utils/sendMail.js';
+import path from 'node:path';
+import fs from 'node:fs/promises';
+import handlebars from 'handlebars';
 
-export const registerUser = async (userData) => {
-  const user = await UsersCollection.findOne({ email: userData.email });
-  if (user) {
-    throw createHttpError(409, 'Email in use');
-  }
-  const encryptedPassword = await bcrypt.hash(userData.password, 10);
-  const createdUser = await UsersCollection.create({
-    ...userData,
+export const registerUser = async (payload) => {
+  const user = await UsersCollection.findOne({ email: payload.email });
+
+  if (user) throw createHttpError(409, 'Email in use');
+
+  const encryptedPassword = await bcrypt.hash(payload.password, 10);
+
+  return await UsersCollection.create({
+    ...payload,
     password: encryptedPassword,
   });
-
-  return createdUser;
 };
 
-export const loginUser = async (userData) => {
-  const user = await UsersCollection.findOne({ email: userData.email });
+export const loginUser = async (payload) => {
+  const user = await UsersCollection.findOne({ email: payload.email });
 
   if (!user) {
-    throw createHttpError(401, 'Wrong email or password');
+    throw createHttpError(404, 'User not found');
   }
 
-  const isValidPassword = await bcrypt.compare(
-    userData.password,
-    user.password,
-  );
+  const isEqual = await bcrypt.compare(payload.password, user.password);
 
-  if (!isValidPassword) {
-    throw createHttpError(401, 'Wrong email or password');
+  if (!isEqual) {
+    throw createHttpError(401, 'Unauthorized');
   }
 
   await SessionsCollection.deleteOne({ userId: user._id });
@@ -42,18 +47,32 @@ export const loginUser = async (userData) => {
   const accessToken = randomBytes(30).toString('base64');
   const refreshToken = randomBytes(30).toString('base64');
 
-  const session = await SessionsCollection.create({
+  return await SessionsCollection.create({
     userId: user._id,
     accessToken,
     refreshToken,
     accessTokenValidUntil: new Date(Date.now() + FIFTEEN_MINUTES),
     refreshTokenValidUntil: new Date(Date.now() + THIRTY_DAYS),
   });
-
-  return session;
 };
 
-export const refreshUserSession = async ({ refreshToken, sessionId }) => {
+export const logoutUser = async (sessionId) => {
+  await SessionsCollection.deleteOne({ _id: sessionId });
+};
+
+const createSession = () => {
+  const accessToken = randomBytes(30).toString('base64');
+  const refreshToken = randomBytes(30).toString('base64');
+
+  return {
+    accessToken,
+    refreshToken,
+    accessTokenValidUntil: new Date(Date.now() + FIFTEEN_MINUTES),
+    refreshTokenValidUntil: new Date(Date.now() + THIRTY_DAYS),
+  };
+};
+
+export const refreshUsersSession = async ({ sessionId, refreshToken }) => {
   const session = await SessionsCollection.findOne({
     _id: sessionId,
     refreshToken,
@@ -63,32 +82,89 @@ export const refreshUserSession = async ({ refreshToken, sessionId }) => {
     throw createHttpError(401, 'Session not found');
   }
 
-  const isRefreshTokenExpired =
+  const isSessionTokenExpired =
     new Date() > new Date(session.refreshTokenValidUntil);
 
-  if (isRefreshTokenExpired) {
+  if (isSessionTokenExpired) {
     throw createHttpError(401, 'Session token expired');
   }
 
-  await SessionsCollection.deleteOne({
-    _id: sessionId,
-    refreshToken,
-  });
+  const newSession = createSession();
 
-  const newAccessToken = randomBytes(30).toString('base64');
-  const newRefreshToken = randomBytes(30).toString('base64');
+  await SessionsCollection.deleteOne({ _id: sessionId, refreshToken });
 
-  const newSession = await SessionsCollection.create({
+  return await SessionsCollection.create({
     userId: session.userId,
-    accessToken: newAccessToken,
-    refreshToken: newRefreshToken,
-    accessTokenValidUntil: new Date(Date.now() + FIFTEEN_MINUTES),
-    refreshTokenValidUntil: new Date(Date.now() + THIRTY_DAYS),
+    ...newSession,
   });
-
-  return newSession;
 };
 
-export const logoutUser = async (sessionId) => {
-  await SessionsCollection.deleteOne({ _id: sessionId });
+export const requestResetToken = async (email) => {
+  const user = await UsersCollection.findOne({ email });
+
+  if (!user) {
+    throw createHttpError(404, 'User not found');
+  }
+
+  const resetToken = jwt.sign(
+    { sub: user._id, email },
+    getEnvVar('JWT_SECRET'),
+    { expiresIn: '5m' },
+  );
+
+  const resetPasswordTemplatePath = path.join(
+    TEMPLATES_DIR,
+    'reset-password-email.html',
+  );
+
+  const templateSource = (
+    await fs.readFile(resetPasswordTemplatePath)
+  ).toString();
+
+  const template = handlebars.compile(templateSource);
+
+  const html = template({
+    name: user.name,
+    link: `${getEnvVar('APP_DOMAIN')}/auth/reset-pwd?token=${resetToken}`,
+  });
+
+  try {
+    await sendEmail({
+      from: getEnvVar(SMTP.SMTP_FROM),
+      to: email,
+      subject: 'Reset your password',
+      html,
+    });
+  } catch {
+    throw createHttpError(
+      500,
+      'Failed to send the email, please try again later.',
+    );
+  }
+};
+
+export const resetPassword = async (payload) => {
+  let entries;
+
+  try {
+    entries = jwt.verify(payload.token, getEnvVar('JWT_SECRET'));
+  } catch {
+    throw createHttpError(401, 'Token is expired or invalid.');
+  }
+
+  const user = await UsersCollection.findOne({
+    email: entries.email,
+    _id: entries.sub,
+  });
+
+  if (!user) throw createHttpError(404, 'User not found');
+
+  const encryptedPassword = await bcrypt.hash(payload.password, 10);
+
+  await SessionsCollection.deleteOne({ userId: user._id });
+
+  await UsersCollection.updateOne(
+    { _id: user._id },
+    { password: encryptedPassword },
+  );
 };
